feat: クライアントサイドでの認証機能実装
