+++ auth.ts
@@ -1,5 +1,6 @@
 import NextAuth from "next-auth";
+import Google from "next-auth/providers/google";
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
-    providers: [],
+    providers: [ Google ],
 })
